+++ identity-iota/src/tangle/client_builder.rs
@@ -33,6 +33,7 @@ impl ClientBuilder {
   }
 
   /// Sets the IOTA Tangle network.
+  #[must_use]
   pub fn network(mut self, network: Network) -> Self {
     self.builder = self.builder.with_network(network.name_str());
     self.network = network;
@@ -40,6 +41,7 @@ impl ClientBuilder {
   }
 
   /// Sets the DID message encoding used when publishing to the Tangle.
+  #[must_use]
   pub fn encoding(mut self, encoding: DIDMessageEncoding) -> Self {
     self.encoding = encoding;
     self
@@ -86,12 +88,14 @@ impl ClientBuilder {
   }
 
   /// Sets the node sync interval.
+  #[must_use]
   pub fn node_sync_interval(mut self, value: Duration) -> Self {
     self.builder = self.builder.with_node_sync_interval(value);
     self
   }
 
   /// Disables the node sync process.
+  #[must_use]
   pub fn node_sync_disabled(mut self) -> Self {
     self.builder = self.builder.with_node_sync_disabled();
     self
@@ -105,36 +109,44 @@ impl ClientBuilder {
   }
 
   /// Enables/disables quorum.
+  #[must_use]
   pub fn quorum(mut self, value: bool) -> Self {
     self.builder = self.builder.with_quorum(value);
     self
   }
 
   /// Sets the number of nodes used for quorum.
+  #[must_use]
   pub fn quorum_size(mut self, value: usize) -> Self {
     self.builder = self.builder.with_quorum_size(value);
     self
   }
 
   /// Sets the quorum threshold.
+  #[must_use]
   pub fn quorum_threshold(mut self, value: usize) -> Self {
     self.builder = self.builder.with_quorum_threshold(value);
     self
   }
 
-  /// Sets whether PoW is performed locally or remotely.
+  /// Sets whether proof-of-work (PoW) is performed locally or remotely.
+  ///
+  /// NOTE: the client will always fallback to local PoW if no nodes support remote PoW.
+  #[must_use]
   pub fn local_pow(mut self, value: bool) -> Self {
     self.builder = self.builder.with_local_pow(value);
     self
   }
 
   /// Sets the number of seconds that new tips will be requested during PoW.
+  #[must_use]
   pub fn tips_interval(mut self, value: u64) -> Self {
     self.builder = self.builder.with_tips_interval(value);
     self
   }
 
   /// Sets the default request timeout.
+  #[must_use]
   pub fn request_timeout(mut self, value: Duration) -> Self {
     self.builder = self.builder.with_request_timeout(value);
     self