+++ CHANGELOG.md
@@ -1,8 +1,27 @@
 # Changelog
 
+## [v0.5.0-dev.2](https://github.com/iotaledger/identity.rs/tree/v0.5.0-dev.2) (2022-01-14)
+
+[Full Changelog](https://github.com/iotaledger/identity.rs/compare/v0.5.0-dev.1...v0.5.0-dev.2)
+
+Changed
+
+- Replace `ClientMap` with `Client` in `Account` [\#582](https://github.com/iotaledger/identity.rs/pull/582)
+- Add signature `created`, `expires`, `challenge`, `domain`, `purpose` [\#548](https://github.com/iotaledger/identity.rs/pull/548)
+
+Added
+
+- Add account synchronization method [\#544](https://github.com/iotaledger/identity.rs/pull/544)
+
+Patch
+
+- Enable local proof-of-work fallback [\#579](https://github.com/iotaledger/identity.rs/pull/579)
+
 ## [v0.5.0-dev.1](https://github.com/iotaledger/identity.rs/tree/v0.5.0-dev.1) (2021-12-15)
 
-[Full Changelog](https://github.com/iotaledger/identity.rs/compare/v0.3.0...v0.5.0-dev.1)
+[Full Changelog](https://github.com/iotaledger/identity.rs/compare/v0.4.0...v0.5.0-dev.1)
+ 
+This release introduces multiple breaking changes to the structure of IOTA DID Documents and their Tangle messages, rendering any identity created with a prior version incompatible and unresolvable. A versioning system has been introduced so any new identities should hopefully be forward compatible with any future breaking changes to the message structure.
 
 Changed
 
@@ -24,6 +43,7 @@ Changed
 Added
 
 - Filter out DiffMessages updating signing methods [\#519](https://github.com/iotaledger/identity.rs/pull/519)
+- Add publish with retry method [\#455](https://github.com/iotaledger/identity.rs/pull/455)
 
 Patch
 
@@ -31,10 +51,17 @@ Patch
 - Improve client error messages [\#512](https://github.com/iotaledger/identity.rs/pull/512)
 - Make `create_signature` and `sign` async for `RemoteEd25519` [\#491](https://github.com/iotaledger/identity.rs/pull/491)
 - Fix credential validation failing for documents with diff updates [\#490](https://github.com/iotaledger/identity.rs/pull/490)
+- Upgrade to the Rust 2021 edition [\#449](https://github.com/iotaledger/identity.rs/pull/449)
+
+## [v0.4.0](https://github.com/iotaledger/identity.rs/tree/v0.4.0) (2021-11-01)
+
+[Full Changelog](https://github.com/iotaledger/identity.rs/compare/v0.3.0...v0.4.0)
 
 ## [v0.3.0](https://github.com/iotaledger/identity.rs/tree/v0.3.0) (2021-05-10)
 
 [Full Changelog](https://github.com/iotaledger/identity.rs/compare/v0.2.0...v0.3.0)
+ 
+This release introduces the high-level `Account` API for creating and managing IOTA identities.
 
 ## [v0.2.0](https://github.com/iotaledger/identity.rs/tree/v0.2.0) (2021-02-18)